Patch Friendly PhysicalCombatOverhaul
I was tired of completely rebuilding this mod so I could patch it for my mods, so, I went through and change all properties in the namespace to public to allow access through namespace use.

With this added, I could easily patch the specific formula without rebuilding the whole mod by calling the namespace properties that are now public.