Magical Compass Effects & Glass Update
I forgot to push the update for this. Here it is. However, the installer rars are to large to push with the new image files. Until I can optimize these effect textures, you will have to either make your own rar by compiling the folder yourself or just use the folder and copy and paste.