Major Optimization Update
With the assist of King of Worms, the canvas objects have been completely rebuilt to properly optimize texture use. This speeds up load time for everything, especially when loading a new game. However, this also helps when loading new cities, since the render texture is a optimized and scaled size.